refactor: rename features/mcp -> features/liferay-mcp (naming collision fix)
