+++ src/commands/mcp/mcp.command.ts
@@ -8,7 +8,7 @@ import {
   runMcpCheck,
   runMcpOpenApis,
   runMcpProbe,
-} from '../../features/mcp/mcp.js';
+} from '../../features/liferay-mcp/liferay-mcp.js';
 import {formatMcpDoctor, runMcpDoctor} from '../../features/mcp-server/mcp-server-doctor.js';
 import type {McpTool} from '../../features/mcp-server/mcp-server-setup.js';
